Separate global_planner_node executable
